+++ lib/copperegg/ver.rb
@@ -1,3 +1,3 @@
 module CopperEgg
-  GEM_VERSION = '0.6.0.pre'
+  GEM_VERSION = '0.6.0.pre2'
 end